requirements.txt
python-telegram-bot==13.15
pyTelegramBotAPI==4.12.0
requests==2.28.2
pytz==2023.3
schedule==1.2.0
psutil==5.9.5
python-dotenv==1.0.0 